+++ node/tests/api/asset/upload.test.ts
@@ -112,7 +112,7 @@ describe.concurrent(pathname, async () => {
                 test: async (response, payload) => {
                     it("test the result of updating file", async () => {
                         /* 测试响应体中的结果 */
-                        payload.files
+                        payload!.files
                             .map((file) => file.name)
                             .forEach((filename) => {
                                 expect.soft(response.data.succMap[filename], `uploaded file "${filename}" failed`).not.toBeUndefined();